# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Tiffin & Student Analytics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meals:
    get:
      operationId: listMeals
      tags: [meals]
      summary: List meal records for the student
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Meal records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealRecord"
    post:
      operationId: createMeal
      tags: [meals]
      summary: Create or update a meal record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealInput"
      responses:
        "201":
          description: Meal record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealRecord"
  /meals/summary:
    get:
      operationId: getMealSummary
      tags: [meals]
      summary: Get meal summary stats
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Meal summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealSummary"

  # ── STUDENT ANALYTICS ─────────────────────────────────────────────────────
...
  /owners/profile:
    get:
      operationId: getOwnerProfile
      tags: [owners]
      summary: Get owner business profile
      responses:
        "200":
          description: Owner profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerProfile"
    put:
      operationId: updateOwnerProfile
      tags: [owners]
      summary: Update owner business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerProfile"
...
  /analytics/owner/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [owner-analytics]
      summary: Get revenue analytics (daily/weekly/monthly)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
        - name: months
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Revenue analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueAnalytics"
...
components:
  schemas:
...
    MealRecord:
      type: object
      required: [id, studentId, date, morningMeal, eveningMeal]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        date:
          type: string
          format: date
        morningMeal:
          type: boolean
        eveningMeal:
          type: boolean
        notes:
          type: ["string", "null"]
    MealInput:
      type: object
      required: [date, morningMeal, eveningMeal]
      properties:
        date:
          type: string
          format: date
        morningMeal:
          type: boolean
        eveningMeal:
          type: boolean
        notes:
          type: string
    MealSummary:
      type: object
      required: [totalMorning, totalEvening, totalMeals, daysTracked]
      properties:
        totalMorning:
          type: integer
        totalEvening:
          type: integer
        totalMeals:
          type: integer
        daysTracked:
          type: integer
        averageMealsPerDay:
          type: number

    # ── STUDENT ANALYTICS SCHEMAS ─────────────────────────────────────────
...
    TrendPoint:
      type: object
      required: [label, amount]
      properties:
        label:
          type: string
        amount:
          type: number
        date:
          type: ["string", "null"]
...
    OwnerProfile:
      type: object
      required: [id, userId, businessName, ownerName, email]
      properties:
        id:
          type: integer
        userId:
          type: integer
        businessName:
          type: string
        ownerName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        serviceCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    OwnerProfileInput:
      type: object
      properties:
        businessName:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        address:
          type: string
...
    RevenueAnalytics:
      type: object
      required: [period, data, totalRevenue, growth]
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        totalRevenue:
          type: number
        growth:
          type: number
...