# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Tiffin & Student Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  label: string;
  amount: number;
  /** @nullable */
  date?: string | null;
}